NPC states created
I've not got the simplified NPC working with it's state types. continued to be simplified.  It has a default state in which it rotates around to face the player provided they are in view. The second state moves the camera closer and over the shoulder, though this needs a little updating as I feel it is too close, and too sudden