Implemented FINAL SpellWheel
Final version of SpellWheel now up and running.

Required quite some special care as the final image is not a perfect circle but it's stretched. Plus the text slots had to become "responsive" depending on their position within the wheel and many of the final assets had to be perfectly aligned.

Added all final assets to the wheel.